vllm_wiki_generator/requirements.txt
fastapi
uvicorn
tiktoken
python-dotenv
pydantic
faiss-cpu
langchain
openai
jinaai
